requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==3.2.18
django-cors-headers==4.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
PyJWT==2.6.0
pytz==2023.3
requests==2.29.0
sqlparse==0.4.4
urllib3==1.26.15
